upload data

+++ resources/views/dashboard/pages/lifting/examination/view.blade.php
@@ -25,13 +25,19 @@
                                                     href="{{ route('examination.reports.create', ['type' => $type]) }}">{{ trans('Dashboard/Lifting/examination.addReport' . $type) }}</a>
                                             </div>
                                         </div>
-                                        <div class="col-sm-auto">
-                                            <x-dashboard.layouts.delete-selected :route="route('reports.deleteAll')"
-                                                :model="$data"></x-dashboard.layouts.delete-selected>
-                                        </div>
+                                        @if (auth()->user()->can('editor'))
+                                            <div class="col-sm-auto">
+                                                <x-dashboard.layouts.delete-selected :route="route('reports.deleteAll')"
+                                                    :model="$data"></x-dashboard.layouts.delete-selected>
+                                            </div>
+                                        @endif
                                         <div class="col-sm-auto">
                                             <x-dashboard.layouts.download-selected :route="route('reports.downloadAll')" :model="$data"
                                                 :pdfView='$pdfView'></x-dashboard.layouts.download-selected>
+                                        </div>
+                                        <div class="col-sm-auto">
+                                            <x-dashboard.layouts.submit :model="$data"
+                                                :orders="$orders"></x-dashboard.layouts.submit>
                                         </div>
                                     </div>
                                 </div>
@@ -55,6 +61,7 @@
                                                     {{ trans('Dashboard/Lifting/examination.report_number') }}</th>
                                                 <th class="" data-sort="">
                                                     {{ trans('Dashboard/Lifting/examination.exam_date') }}</th>
+                                                <th class="" data-sort="">Job Status</th>
                                                 <th class="" data-sort="">
                                                     {{ trans('Dashboard/Lifting/examination.user') }}</th>
                                                 {{-- <th class="" data-sort="">{{ trans('Dashboard/Lifting/examination.next_date') }}</th> --}}
@@ -85,6 +92,12 @@
                                                     <td class="id">{{ $i++ }}</td>
                                                     <td class="customer_name ">{{ $item->report_number }}</td>
                                                     <td class="customer_full_name">{{ $item->date }}</td>
+                                                    <td class="status">
+                                                        <span
+                                                            class="badge bg-pill @if ($item->status === 'Open') bg-success @else bg-danger @endif ">
+                                                            {{ $item->status }}
+                                                        </span>
+                                                    </td>
                                                     <td class="customer_full_name">{{ $item->getUser->full_name }}</td>
                                                     {{-- <td class="customer_full_name">{{$item->next_exam}}</td> --}}
                                                     @if ($type == 'thorough')
@@ -112,7 +125,7 @@
                                                                         href="{{ route('examination.reports.repeat', $item->id) }}"><i
                                                                             class=" bx bx-repeat me-1"></i>
                                                                         Repeat</a></li>
-                                                                @if (auth()->user()->id == $item->user_id || auth()->user()->can('editor'))
+                                                                @if ((auth()->user()->id == $item->user_id && $item->status == 'Open') || auth()->user()->can('editor'))
                                                                     <li><a class="dropdown-item edit-item-btn"
                                                                             target="_blank"
                                                                             href="{{ route('examination.reports.edit', $item->id) }}"><i
@@ -168,9 +181,9 @@
                                     </table>
                                 </div>
 
-                                <div class="d-flex justify-content-end">
+                                {{-- <div class="d-flex justify-content-end">
                                     {{ $data->links('pagination::bootstrap-5') }}
-                                </div>
+                                </div> --}}
                             </div>
                         </div><!-- end card -->
                     </div>
