+++ resources/views/dashboard/pages/lifting/mpi/view.blade.php
@@ -45,13 +45,19 @@
                                                     href="{{ route('mpi.reports.create') }}">{{ TranslationHelper::translate(ucwords('new mpi')) }}</a>
                                             </div>
                                         </div>
-                                        <div class="col-sm-auto">
-                                            <x-dashboard.layouts.delete-selected :route="route('reports.deleteAll')"
-                                                :model="$data"></x-dashboard.layouts.delete-selected>
-                                        </div>
+                                        @if (auth()->user()->can('editor'))
+                                            <div class="col-sm-auto">
+                                                <x-dashboard.layouts.delete-selected :route="route('reports.deleteAll')"
+                                                    :model="$data"></x-dashboard.layouts.delete-selected>
+                                            </div>
+                                        @endif
                                         <div class="col-sm-auto">
                                             <x-dashboard.layouts.download-selected :route="route('reports.downloadAll')" :model="$data"
                                                 :pdfView='$pdfView'></x-dashboard.layouts.download-selected>
+                                        </div>
+                                        <div class="col-sm-auto">
+                                            <x-dashboard.layouts.submit :model="$data"
+                                                :orders="$orders"></x-dashboard.layouts.submit>
                                         </div>
                                     </div>
                                 </div>
@@ -74,6 +80,7 @@
                                                     {{ trans('Dashboard/Lifting/mpi.report_number') }}
                                                 <th class="" data-sort="">
                                                     {{ trans('Dashboard/Lifting/mpi.date') }}</th>
+                                                <th class="" data-sort="">Job Status</th>
                                                 <th class="" data-sort="">
                                                     {{ trans('Dashboard/Lifting/mpi.user') }}</th>
                                                 <th class="" data-sort="">
@@ -102,6 +109,12 @@
                                                     <td class="id">{{ $i++ }}</td>
                                                     <td class="customer_full_name">{{ $item->report_number }}</td>
                                                     <td class="customer_full_name">{{ $item->date }}</td>
+                                                    <td class="status">
+                                                        <span
+                                                            class="badge bg-pill @if ($item->status === 'Open') bg-success @else bg-danger @endif ">
+                                                            {{ $item->status }}
+                                                        </span>
+                                                    </td>
                                                     <td class="customer_full_name">{{ $item->getUser->name }}</td>
                                                     <td class="customer_full_name">{{ $item->serial }}</td>
                                                     @if ($item->acceptance == 1)
@@ -125,7 +138,7 @@
                                                                         href="{{ route('mpi.reports.repeat', $item->id) }}"><i
                                                                             class=" bx bx-repeat me-1"></i>
                                                                         Repeat</a></li>
-                                                                @if (auth()->user()->id == $item->user_id || auth()->user()->can('editor'))
+                                                                @if ((auth()->user()->id == $item->user_id && $item->status == 'Open') || auth()->user()->can('editor'))
                                                                     <li><a class="dropdown-item edit-item-btn"
                                                                             target="_blank"
                                                                             href="{{ route('mpi.reports.edit', $item->id) }}"><i
@@ -189,9 +202,9 @@
                                         </tbody>
                                     </table>
                                 </div>
-                                <div class="d-flex justify-content-end">
+                                {{-- <div class="d-flex justify-content-end">
                                     {{ $data->links('pagination::bootstrap-5') }}
-                                </div>
+                                </div> --}}
                             </div>
                         </div><!-- end card -->
                     </div>
