+++ resources/views/dashboard/pages/tublar/tubs/view.blade.php
@@ -37,9 +37,19 @@
                                                     href="{{ route('tubs.reports.create', ['type' => $type]) }}">{{ TranslationHelper::translate('new' . ' ' . $type) }}</a>
                                             </div>
                                         </div>
+                                        @if (auth()->user()->can('editor'))
+                                            <div class="col-sm-auto">
+                                                <x-dashboard.layouts.delete-selected :route="route('reports.deleteAll')"
+                                                    :model="$data"></x-dashboard.layouts.delete-selected>
+                                            </div>
+                                        @endif
                                         <div class="col-sm-auto">
-                                            <x-dashboard.layouts.delete-selected :route="route('reports.deleteAll')"
-                                                :model="$data"></x-dashboard.layouts.delete-selected>
+                                            <x-dashboard.layouts.download-selected :route="route('reports.downloadAll')" :model="$data"
+                                                :pdfView='$pdfView'></x-dashboard.layouts.download-selected>
+                                        </div>
+                                        <div class="col-sm-auto">
+                                            <x-dashboard.layouts.submit :model="$data"
+                                                :orders="$orders"></x-dashboard.layouts.submit>
                                         </div>
                                     </div>
                                 </div>
@@ -62,6 +72,7 @@
                                                     {{ TranslationHelper::translate('report_number') }}</th>
                                                 <th class="" data-sort="">
                                                     {{ TranslationHelper::translate('report_date') }}</th>
+                                                <th class="" data-sort="">Job Status</th>
                                                 <th class="" data-sort="">
                                                     {{ TranslationHelper::translate('customer') }}</th>
                                                 <th class="" data-sort="">
@@ -95,6 +106,12 @@
                                                     <td class="id">{{ $i++ }}</td>
                                                     <td class="customer_name ">{{ $item->report_number }}</td>
                                                     <td class="customer_full_name">{{ $item->date }}</td>
+                                                    <td class="status">
+                                                        <span
+                                                            class="badge bg-pill @if ($item->status === 'Open') bg-success @else bg-danger @endif ">
+                                                            {{ $item->status }}
+                                                        </span>
+                                                    </td>
                                                     <td class="customer_full_name">{{ $item->customer }}</td>
                                                     <td class="customer_full_name">{{ $item->getUser->full_name }}</td>
                                                     @if (isset($item->getAccept->value))
@@ -125,7 +142,7 @@
                                                                         href="{{ route('tubs.reports.repeat', $item->id) }}"><i
                                                                             class=" bx bx-repeat me-1"></i>
                                                                         Repeat</a></li>
-                                                                @if (auth()->user()->id == $item->user_id || auth()->user()->can('editor'))
+                                                                @if ((auth()->user()->id == $item->user_id && $item->status == 'Open') || auth()->user()->can('editor'))
                                                                     <li><a class="dropdown-item edit-item-btn"
                                                                             target="_blank"
                                                                             href="{{ route('tubs.reports.edit', $item->id) }}"><i
@@ -194,9 +211,9 @@
                 </tbody>
                 </table>
             </div>
-            <div class="d-flex justify-content-end">
+            {{-- <div class="d-flex justify-content-end">
                 {{ $data->links('pagination::bootstrap-5') }}
-            </div>
+            </div> --}}
         </div>
     </div><!-- end card -->
     </div>
